♻️Add type to user API
