+++ src/types/index.ts
@@ -3,5 +3,6 @@ export * from './core/lecture';
 export * from './core/timetable';
 export * from './core/query';
 export * from './core/notification';
+export * from './core/user';
 
 export * from './api';
